NP-14149 Update types
