+++ src/types/publication_types/reportRegistration.types.ts
@@ -1,7 +1,12 @@
-import { BaseEntityDescription, BaseReference, BaseRegistration } from '../registration.types';
+import {
+  BaseEntityDescription,
+  BaseReference,
+  BaseRegistration,
+  ContextPublisher,
+  Series,
+} from '../registration.types';
 import { PublicationType, ReportType } from '../publicationFieldNames';
 import { PagesMonograph, emptyPagesMonograph } from './pages.types';
-import { ContextPublisher, Series } from './bookRegistration.types';
 
 export interface ReportRegistration extends BaseRegistration {
   entityDescription: ReportEntityDescription;
