+++ src/types/publication_types/researchDataRegistration.types.ts
@@ -1,5 +1,11 @@
 import { PublicationType, ResearchDataType } from '../publicationFieldNames';
-import { BaseEntityDescription, BaseReference, BaseRegistration } from '../registration.types';
+import {
+  BaseEntityDescription,
+  BaseReference,
+  BaseRegistration,
+  ContextPublisher,
+  emptyContextPublisher,
+} from '../registration.types';
 
 export interface ResearchDataRegistration extends BaseRegistration {
   entityDescription: MediaContributionEntityDescription;
@@ -7,18 +13,22 @@ export interface ResearchDataRegistration extends BaseRegistration {
 
 export interface ResearchDataPublicationInstance {
   type: ResearchDataType | '';
+  related: string[];
 }
 
 export const emptyResearchDataPublicationInstance: ResearchDataPublicationInstance = {
   type: '',
+  related: [],
 };
 
 export const emptyResearchDataPublicationContext: ResearchDataPublicationContext = {
   type: PublicationType.ResearchData,
+  publisher: emptyContextPublisher,
 };
 
 export interface ResearchDataPublicationContext {
   type: PublicationType.ResearchData;
+  publisher: ContextPublisher;
 }
 
 interface ResearchDataContributionReference extends BaseReference {
